+++ src/main/java/com/auction/shared/model/user/Bidder.java
@@ -1,11 +1,11 @@
 package com.auction.shared.model.user;
 
-import com.auction.shared.model.auction.Auction;
+import java.math.BigDecimal;
 
 public class Bidder extends User {
     private double balance;
 
-    public Bidder(String username, String password, String fullname, String email, double balance) {
+    public Bidder(String username, String password, String fullname, String email, BigDecimal balance) {
         super(username, password, fullname, email, "BIDDER");
         this.balance = balance;
     }
